+++ test/html/index.spec.ts
@@ -1,0 +1,7 @@
+import iconSpec from './icon.spec';
+import textSpec from './text.spec';
+
+describe('html', () => {
+  iconSpec();
+  textSpec();
+});
